[accel-web] mockSession for test
